Move concepts into concepts
Define type traits there as well, then specialize them at the
top of the resource files based on forward declarations.
This allows them to be used in the implementations of the
types and in Event/Resource.
